Move sendAlert to Utils
