+++ src/main/kotlin/com/kotlindiscord/bot/filtering/InviteFilter.kt
@@ -10,6 +10,7 @@ import com.gitlab.kordlib.core.entity.channel.Channel
 import com.gitlab.kordlib.core.event.message.MessageCreateEvent
 import com.gitlab.kordlib.core.event.message.MessageUpdateEvent
 import com.kotlindiscord.bot.deleteIgnoringNotFound
+import com.kotlindiscord.bot.sendAlert
 import com.kotlindiscord.kord.extensions.ExtensibleBot
 import mu.KotlinLogging
